fix: Address minor bugs
- (PersonalSummary): Added a missing space to the regex for first name,
last name, and profession inputs, allowing spaces in these fields. Unit
tests have been updated to reflect this change.
- (ContactDetails): Removed an erroneous space from the character range
for digits in the mobile number regex, which caused validation errors
for valid inputs.
- (App): Some descriptions inside formDetails made use of arrays to
concatenate strings with HTML elements. This caused issues when
rendering as React expects list items to be assigned key values.
Removed the array and replaced with a jsx element to achieve the same
result.